build: resolve compiler errors/warnings on openBSD/clang
lib/exmdb_client2.cpp:584:30: warning: lambda capture 'now' is not used [-Wunused-lambda-capture]
  584 | std::erase_if(name_to_srv, [now](const decltype(name_to_srv)::value_type &entry) {
tests/exrpctest.cpp:89:3: error: use of undeclared identifier 'sleep'
   89 | sleep(waitx);
exch/ews/context.cpp:2330:39: warning: suggest braces around initialization of subobject [-Wmissing-braces]
 2330 | const TAGGED_PROPVAL rtf_prop[] = {PR_RTF_IN_SYNC, &rtf_nosync};
exch/ews/context.cpp:2487:20: warning: suggest braces around initialization of subobject [-Wmissing-braces]
 2487 | BINARY att_enc{0, nullptr};
exch/ews/context.cpp:4063:50: warning: suggest braces around initialization of subobject [-Wmissing-braces]
 4063 | static_cast<uint32_t>(ep.m_offset), ep.m_udata});
exch/ews/context.cpp:4071:50: warning: suggest braces around initialization of subobject [-Wmissing-braces]
 4071 | static_cast<uint32_t>(ep.m_offset), ep.m_udata});
exch/ews/structures.cpp:2449:15: warning: suggest braces around initialization of subobject [-Wmissing-braces]
 2449 | static_cast<uint8_t *>(EWSContext::alloc(bin.size()))});

+++ exch/oab.cpp
@@ -1463,8 +1463,9 @@ bool OabPlugin::generate_oab(int32_t base_id, oab_cache_entry &entry)
 	auto full = doc.NewElement("Full");
 	full->SetAttribute("seq", entry.sequence);
 	full->SetAttribute("ver", OAB_V4_VERSION);
-	full->SetAttribute("size", entry.lzx_data.size());
-	full->SetAttribute("uncompressedsize", raw.size());
+	/* ambiguity warning in clang-19 warrants static_cast */
+	full->SetAttribute("size", static_cast<uint64_t>(entry.lzx_data.size()));
+	full->SetAttribute("uncompressedsize", static_cast<uint64_t>(raw.size()));
 	full->SetAttribute("SHA", data_sha.c_str());
 	full->SetText((std::to_string(entry.sequence) + ".lzx").c_str());
 	oal->InsertEndChild(full);
@@ -1472,8 +1473,8 @@ bool OabPlugin::generate_oab(int32_t base_id, oab_cache_entry &entry)
 	auto tmpl = doc.NewElement("Template");
 	tmpl->SetAttribute("seq", entry.sequence);
 	tmpl->SetAttribute("ver", OAB_TMPL_VERSION);
-	tmpl->SetAttribute("size", entry.tmpl_lzx_data.size());
-	tmpl->SetAttribute("uncompressedsize", tmpl_raw.size());
+	tmpl->SetAttribute("size", static_cast<uint64_t>(entry.tmpl_lzx_data.size()));
+	tmpl->SetAttribute("uncompressedsize", static_cast<uint64_t>(tmpl_raw.size()));
 	tmpl->SetAttribute("SHA", tmpl_sha.c_str());
 	tmpl->SetAttribute("langid", "0409");
 	tmpl->SetAttribute("type", "windows");
